Finished integrationwith database in GraphQL, started frontend application

+++ projeto-04/backend/resolvers/Mutation/usuario.js
@@ -13,7 +13,9 @@ const mutations = {
             }
         })
     },
-    async novoUsuario(_, { dados }) {
+    async novoUsuario(_, { dados }, ctx) {
+        if(ctx)ctx.validarAdmin()
+
         try {
             const idsPerfis = []
 
@@ -49,7 +51,9 @@ const mutations = {
             throw new Error(e)
         }
     },
-    async excluirUsuario(_, args) {
+    async excluirUsuario(_, args, ctx) {
+        if(ctx)ctx.validarAdmin()
+
         try {
             const usuario = await obterUsuario(_, args)
             if(usuario) {
@@ -65,12 +69,13 @@ const mutations = {
         }
 
     },
-    async alterarUsuario(_, { filtro, dados }) {
+    async alterarUsuario(_, { filtro, dados }, ctx) {
+        ctx && ctx.validarUsuarioFiltro(filtro)
         try {
             const usuario = await obterUsuario(_, { filtro })
             if(usuario) {
                 const { id } = usuario
-                if(dados.perfis) {
+                if(ctx.admin && dados.perfis) {
                     await db('usuarios_perfis')
                         .where({ usuario_id: id }).delete()
